pyproject.toml
[project]
name = "autoinstallmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
